feat: adds _generate_leds to default_header.h
This feature makes it possible for users to reference an LED by key.
This requires a label property on each key along with the matrix in
info.json.

    rgb_matrix_set_color(LED_INDEX(Q),    SET_COLOR);

It requires a simple define.  Layout must be known.

        g_led_config.matrix_co[LED_LAYOUT_65_ansi_blocker_##KEY##_ROW] \
                              [LED_LAYOUT_65_ansi_blocker_##KEY##_COL]

Some may say that referencing a key by label assume an assignment, but
there is no better way of referencing a key.  Without this compromise,
developers would have to manually look up the LED index according to
the row/column in the matrix property in info.js, for each key's LED
they want to utilize.